组织机构 功能模块重构

+++ myclouds-system-service/src/main/java/com/gitee/myclouds/system/module/org/OrgService.java
@@ -6,11 +6,14 @@ import org.apache.ibatis.session.SqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.alibaba.fastjson.JSON;
-import com.gitee.myclouds.base.vo.UserVO;
+import com.gitee.myclouds.base.exception.BizException;
+import com.gitee.myclouds.base.helper.treebuiler.TreeBuilder;
+import com.gitee.myclouds.base.helper.treebuiler.TreeNodeVO;
+import com.gitee.myclouds.base.vo.OutVO;
 import com.gitee.myclouds.common.util.MyUtil;
 import com.gitee.myclouds.common.wrapper.Dto;
 import com.gitee.myclouds.common.wrapper.Dtos;
+import com.gitee.myclouds.system.domain.mymodule.MyModuleEntity;
 import com.gitee.myclouds.system.domain.myorg.MyOrgEntity;
 import com.gitee.myclouds.system.domain.myorg.MyOrgMapper;
 
@@ -22,73 +25,77 @@ import com.gitee.myclouds.system.domain.myorg.MyOrgMapper;
  */
 @Service
 public class OrgService {
-	
+
 	@Autowired
 	private SqlSession sqlSession;
 	@Autowired
 	private MyOrgMapper myOrgMapper;
-	
+
 	/**
 	 * 查询列表
 	 * 
 	 * @param inDto
 	 * @return
 	 */
-	public String list(Dto inDto) {
-		Dto outDto = Dtos.newDto();
-		List<Dto> myOrgDtos = sqlSession.selectList("sql.org.pageOrg",inDto);
-		Integer total = sqlSession.selectOne("sql.org.pageOrgCount", inDto);
-		outDto.put("data", myOrgDtos);
-		outDto.put("recordsTotal", total);
-		outDto.put("recordsFiltered", total);
-		return JSON.toJSONString(outDto);
+	public OutVO list(Dto inDto) {
+		OutVO outVO = new OutVO(0);
+		List<MyModuleEntity> myModuleEntitys = sqlSession.selectList("sql.org.listOrg", inDto);
+		outVO.setData(myModuleEntitys).setCount(myModuleEntitys.size());
+		return outVO;
 	}
-	
+
 	/**
 	 * 查询实体
 	 * 
 	 * @param id
 	 * @return
 	 */
-	public String get(Integer id) {
+	public OutVO get(Integer id) {
+		OutVO outVO  = new OutVO(0);
 		MyOrgEntity myOrgEntity = myOrgMapper.selectByKey(id);
-		return JSON.toJSONString(myOrgEntity);
+		MyOrgEntity parentEntity = myOrgMapper.selectByKey(myOrgEntity.getParent_id());
+		Dto dto = Dtos.newDto();
+		MyUtil.copyProperties(myOrgEntity, dto);
+		if (MyUtil.isNotEmpty(parentEntity)) {
+			dto.put("parent_name", parentEntity.getName());
+		}else {
+			dto.put("parent_name", "无");
+		}
+		outVO.setData(dto);
+		return outVO;
 	}
-	
+
 	/**
 	 * 新增
 	 * 
 	 * @param inDto
 	 * @return
 	 */
-	public Dto save(Dto inDto) {
-		Dto outDto = null;
-		// 拷贝参数对象中的属性到实体对象中
+	public OutVO add(Dto inDto) {
+		OutVO outVO = new OutVO(0);
 		MyOrgEntity myOrgEntity = new MyOrgEntity();
 		MyUtil.copyProperties(inDto, myOrgEntity);
-		//TODO
-		UserVO curUser = null;
-		myOrgEntity.setCreate_by(curUser.getName());
-		myOrgEntity.setCreate_by_id(curUser.getId());
+		//TODO curUser
+		myOrgEntity.setCreate_by("超级用户");
+		myOrgEntity.setCreate_by_id(1);
 		myOrgMapper.insert(myOrgEntity);
-		outDto = Dtos.newDto().set("code", "1").set("msg", "部门新增成功");
-		return outDto;
+		outVO.setMsg("组织新增成功");
+		return outVO;
 	}
-	
+
 	/**
 	 * 修改
 	 * 
 	 * @param inDto
 	 * @return
 	 */
-	public Dto update(Dto inDto) {
-		Dto outDto = null;
-		//拷贝参数对象中的属性到实体对象中
+	public OutVO update(Dto inDto) {
+		OutVO outVO  = new OutVO(0);
 		MyOrgEntity myOrgEntity = new MyOrgEntity();
 		MyUtil.copyProperties(inDto, myOrgEntity);
 		myOrgMapper.updateByKey(myOrgEntity);
-		outDto = Dtos.newDto().set("code", "1").set("msg", "部门修改成功");
-		return outDto;
+		outVO.setMsg("组织修改成功");
+		return outVO;
 	}
 
 	/**
@@ -97,30 +104,32 @@ public class OrgService {
 	 * @param inDto
 	 * @return
 	 */
-	public Dto delete(Dto inDto) {
-		Dto outDto = Dtos.newDto();
-		Integer orgId = inDto.getInteger("id");
+	public OutVO delete(Integer orgId) {
+		OutVO outVO = new OutVO(0);
+		if (orgId == 1) {
+			throw new BizException(-13, "操作取消，根组织不能删除");
+		}
 		Integer total = sqlSession.selectOne("sql.org.countOrgs", orgId);
 		if (total > 0) {
-			outDto.set("code", "-1").set("msg", "部门删除失败。请先删除下级部门。");
-		}else {
-			myOrgMapper.deleteByKey(orgId);
-			sqlSession.delete("sql.org.deleteMyUser", orgId);
-			outDto.set("code", "1").set("msg", "部门删除成功");
+			throw new BizException(-14, "操作取消，请先删除子部门");
 		}
-		return outDto;
+		myOrgMapper.deleteByKey(orgId);
+		sqlSession.delete("sql.org.deleteMyUser", orgId);
+		outVO.setMsg("组织删除成功");
+		return outVO;
 	}
 	
 	/**
-	 * 查询部门树
+	 * 查询组织树（返回树状数据模型）
 	 * 
 	 * @param inDto
 	 * @return
 	 */
-	//TODO 没有支持子部门操作，迭代版本支持
-	public String listOrgTree(Dto inDto) {
-
-		return null;
+	public OutVO listOrgTree(Dto inDto) {
+		OutVO outVO  = new OutVO(0);
+		List<TreeNodeVO> treeNodeVOs = sqlSession.selectList("sql.org.listOrgTree", inDto);
+		outVO.setData(new TreeBuilder(treeNodeVOs).buildTree());
+		return outVO;
 	}
-	
+
 }
